# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E2EE Chat API",
        "version": "1.0.0",
        "description": "API documentation for the E2EE Chat application."
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/api/users/me": {
            "get": {
                "summary": "Get current authenticated user",
                "tags": ["User"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "User ID",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid bearer token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/users/available": {
            "get": {
                "summary": "Get list of available users (excluding current user)",
                "tags": ["User"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of available users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "user@example.com"
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid bearer token"
                    }
                }
            }
        },
        "/api/users/search": {
            "get": {
                "summary": "Search user by email",
                "tags": ["User"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        },
                        "description": "Email address to search for",
                        "example": "user@example.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        },
                                        "createdAt": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email parameter is required"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid bearer token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/chats/between": {
            "get": {
                "summary": "Get chat between two users (public endpoint for demonstration)",
                "tags": ["Chat"],
                "security": [],
                "parameters": [
                    {
                        "name": "user1Id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "First user ID"
                    },
                    {
                        "name": "user2Id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Second user ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat between two users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "participants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing or invalid user IDs"
                    },
                    "404": {
                        "description": "Chat not found between these users"
                    }
                }
            }
        },
        "/api/messages/public": {
            "get": {
                "summary": "Get encrypted messages from a chat (public endpoint - messages are encrypted, cannot be read without private keys)",
                "tags": ["Message"],
                "security": [],
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Chat ID to fetch encrypted messages from"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encrypted messages from chat (content is encrypted and unreadable without private key)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "ciphertext": {
                                                "type": "string",
                                                "description": "Encrypted message content (Base64 encoded)"
                                            },
                                            "iv": {
                                                "type": "string",
                                                "description": "Initialization vector for AES-GCM (Base64 encoded)"
                                            },
                                            "senderId": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "sender": {
                                                "type": "object",
                                                "properties": {
                                                    "email": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing chatId parameter"
                    },
                    "404": {
                        "description": "Chat not found"
                    }
                }
            }
        }
    }
}